// components/TestGeneratedComponent/0.1.0.0/src/ts/layer3/TestGeneratedComponentModel.interface.ts
/**
 * TestGeneratedComponentModel - TestGeneratedComponent Component Model Interface
 * Web4 pattern: Component model following auto-discovery patterns
 */

import { Model } from './Model.interface.js';

export interface TestGeneratedComponentModel extends Model {
  uuid: string;
  name: string;
  origin: string;
  definition: string;
  createdAt: string;
  updatedAt: string;
}